refactored homeTabs.jsx, created diagnosticsTab.jsx
